Use required Java versions at build and run time
Buck requires Java 8 to build and Java 11 to run. [1]

This has caused spurious failures in the PKGBUILD [2] because neither
did the PKGBUILD enforce the makedepends version to be actually *used*
in the build, nor did it enforce the runtime Java version.

Currently, there’s no method to pin JDKs/JREs in a vendor-agnostic
way [3]. This commit therefore pins the dependencies to the `-openjdk`
vendor, which is better than nothing. It may cause redundant installs
for people who already have JRE 8 and JDK 11 installed from different
vendors. But that redundancy should be a small price to pay for getting
the issue fixed for everyone.

A similar fix has already worked well in the past [4] on macOS. So it’s
pretty safe to assume the fix is good enough for the PKGBUILD as well.

[1]: https://github.com/facebook/buck/issues/2443

[2]: https://aur.archlinux.org/packages/buck/#comment-738767

[3]: https://wiki.archlinux.org/title/Java#Package_pre-requisites_to_support_archlinux-java

[4]: https://github.com/facebook/homebrew-fb/pull/52

CC: Kenneth Endfinger <kaendfinger@gmail.com>
Reported-by: sramey <sramey40@gmail.com>